[update] : Sidebarからコンポーネントを分割
